requirements.txt
streamlit
python-docx
azure-ai-textanalytics>=5.3.0
docx